requirements.txt
alembic==1.13.1
blinker==1.7.0
certifi==2023.11.17
chardet==3.0.4
click==8.1.7
dominate==2.9.1
Faker==22.2.0
Flask==3.0.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.5.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.4
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.24.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.2
